build: move cacao tests to integration, later split out again
