+++ packages/integration/.eslintrc.json
@@ -1,0 +1,11 @@
+{
+  "extends": ["3box", "3box/jest", "3box/typescript"],
+  "parserOptions": {
+    "project": ["tsconfig.lint.json"]
+  },
+  "rules": {
+    "@typescript-eslint/ban-ts-comment": "off",
+    "@typescript-eslint/ban-ts-ignore": "off",
+    "@typescript-eslint/interface-name-prefix": "off"
+  }
+}
