+++ packages/integration/.gitignore
@@ -1,0 +1,4 @@
+*.log
+.DS_Store
+node_modules
+dist
