+++ packages/integration/README.md
@@ -1,0 +1,11 @@
+# DID + CACAO Integration Tests
+
+## Usage
+
+```
+npm run test
+```
+
+## License
+
+Apache-2.0 OR MIT
