+++ packages/integration/package.json
@@ -1,0 +1,60 @@
+{
+  "name": "integration",
+  "version": "0.0.1",
+  "author": "3Box Labs",
+  "license": "(Apache-2.0 OR MIT)",
+  "type": "module",
+  "main": "dist/index.js", 
+  "types": "dist/index.d.ts",
+  "exports": {
+    ".": "./dist/index.js"
+  },
+  "files": [
+    "dist"
+  ],
+  "engines": {
+    "node": ">=14.14"
+  },
+  "sideEffects": false,
+  "scripts": {
+    "build:clean": "del dist",
+    "build:js": "swc src -d ./dist --config-file ../../.swcrc",
+    "build:types": "tsc --emitDeclarationOnly --skipLibCheck",
+    "build": "pnpm run build:clean && pnpm run build:types && pnpm run build:js",
+    "lint": "eslint src --fix",
+    "test": "node --experimental-vm-modules ../../node_modules/jest/bin/jest.js",
+    "test:ci": "pnpm run test --ci --coverage",
+    "prepare": "pnpm run build",
+    "prepublishOnly": "package-check",
+    "size": "./node_modules/.bin/size-limit",
+    "analyze": "./node_modules/.bin/size-limit --why"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ceramicnetwork/js-did.git"
+  },
+  "keywords": [
+    "DID",
+    "identity",
+    "did-provider",
+    "self-sovereign"
+  ],
+  "bugs": {
+    "url": "https://github.com/ceramicnetwork/js-did/issues"
+  },
+  "homepage": "https://github.com/ceramicnetwork/js-did#readme",
+  "devDependencies": {
+    "ceramic-cacao": "workspace:^1.3.1",
+    "ethereum-cacao": "workspace:^0.0.1",
+    "solana-cacao": "workspace:^0.0.1",
+    "@ethersproject/wallet": "^5.5.0",
+    "@ipld/dag-cbor": "^7.0.1",
+    "@stablelib/ed25519": "^1.0.2",
+    "apg-js": "^4.1.1",
+    "caip": "^1.0.0",
+    "multiformats": "^9.5.1",
+    "uint8arrays": "^3.0.0",
+    "@types/luxon": "^3.0.0",
+    "luxon": "^3.0.1"
+  }
+}
